Map and movement updates
This commit updates the map/board drawing to reflect the layout as presented by the project description. Changes to the movement mechanism now also reflect the non-holonomic differential drive style of movement. Currently unable to figure out how to handle theta changes and their placement in the V matrix. May need to simply stick with X and Y coordinates in the matrix, since potentially all possible orientation angles could be used to reach any given point on the map.