Refined A* Search
This version re-implemented the theta index calculation for the V matrix, and refined it to essentially establish a 30 degree "threshold" for orientation on any node. Utilizes modulus to first convert the theta into a standard 360 degree range, and then a secondary modulus with 30 to determine if the value should be rounded up (15+ degree difference) or down.